ticket:35 Moved jabber code to package
