+++ p2psafety_django/events/jabber/__init__.py
@@ -1,0 +1,8 @@
+import logging
+
+
+logger = logging.getLogger('events.jabber')
+
+
+from .clients import PubsubClient
+from .queries import *
