+++ p2psafety_django/events/jabber/clients.py
@@ -2,18 +2,13 @@ import functools
 import logging
 from contextlib import contextmanager
 
-from django.conf import settings
-from django.contrib.auth.models import User
 from django.core.exceptions import ImproperlyConfigured
 
 from sleekxmpp import ClientXMPP
 from sleekxmpp.xmlstream import ET
 from lxml import etree
 
-import users.utils
-
-
-logger = logging.getLogger('events.jabber')
+from . import logger
 
 
 class PubsubClient(object):
@@ -77,59 +72,3 @@ class PubsubClient(object):
         self._pubsub.publish(self.config.pubsub_server,
                              self.config.node_name,
                              payload=payload)
-
-
-def _get_registered_users():
-    """
-    Returns registered users as list of jid-s.
-    """
-    return []
-
-
-def _create_account(user):
-    jabber_username = '%s@p2psafety.net' % user.username
-    jabber_password = users.utils.get_api_key(user).key
-    return jabber_username
-
-
-def on_user_created(new_user):
-    """
-    This function should be called for newly registered user.
-
-    :type new_user: `django.contrib.auth.models.User`
-    """
-    _create_account(new_user)
-
-
-def on_startup():
-    """
-    This function should be called on server start.
-    """
-    logger.info('jabber account synchronization started')
-    jabber_users = _get_registered_users()
-    jabber_usernames = [jid.split('@')[0] for jid in jabber_users]
-    site_users = User.objects.only('id', 'username')
-    created_jids = [_create_account(u) for u in site_users
-                    if u.username not in jabber_usernames]
-    logger.info('created %d accounts for: %s', len(created_jids), created_jids)
-
-
-def notify_supporters(event):
-    """
-    Sends notifications to event's supporters via jabber node.
-
-    :type event: :class:`events.models.Event`
-    """
-    from events.api.resources import EventResource
-
-    # Constructing payload
-    resource = EventResource()
-    event_dict = resource.full_dehydrate(resource.build_bundle(obj=event))
-    payload = resource.serialize(None, event_dict, 'application/xml')
-
-    with PubsubClient(settings.EVENTS_NOTIFIER) as client:
-        client.publish(payload)
-
-
-def notify_supporter(event, supporter):
-    raise NotImplementedError
